% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_function_sharp_null.R
\name{assign_CRE}
\alias{assign_CRE}
\title{Generate matrix of complete randomization assignments}
\usage{
assign_CRE(n, m, nperm)
}
\arguments{
\item{n}{A positive integer that specifies the number of units.}

\item{m}{A positive integer that specifies the number of treated units.}

\item{nperm}{A positive integer representing the number of permutations.}
}
\value{
An \eqn{n \times nperm} matrix.
}
\description{
Draw multiple assignments from a completely randomized experiment
}
\examples{
assign_CRE(n = 5, m = 3, nperm = 10^4)
}
