% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_function_quantile.R
\name{ci_quantile}
\alias{ci_quantile}
\title{Randomization inference for quantiles of individual treatment effects with missing outcomes}
\usage{
ci_quantile(
  Z,
  Y,
  k.vec = NULL,
  missing = "general",
  MWU = FALSE,
  method.list.treated = list(name = "Wilcoxon"),
  method.list.control = list(name = "Wilcoxon"),
  stat.null.treated = NULL,
  Z.perm.treated = NULL,
  stat.null.control = NULL,
  Z.perm.control = NULL,
  nperm = 10^4,
  alpha = 0.05,
  tol = 10^(-3)
)
}
\arguments{
\item{Z}{Treatment assignment (\eqn{n \times 1} vector).}

\item{Y}{Observed outcome (\eqn{n \times 1} vector, including \code{NA}s).}

\item{k.vec}{A vector that specifies the quantiles of individual effects of interest.
If \code{k.vec = NULL}, then we consider all quantiles of individual effects \eqn{1 \leq k \leq n}.}

\item{missing}{A string specifying the missing mechanism.
\itemize{
\item \code{missing = "general"}: general missing mechanism.
\item \code{missing = "mp"}: monotone positive missing mechanism (\eqn{M_1 >= M_0}).
\item \code{missing = "mn"}: monotone negative missing mechanism (\eqn{M_1 <= M_0}).
\item \code{missing = "sharp"}: sharp missing mechanism (\eqn{M_1 = M_0}).
\item \code{missing = "random"}: random missing mechanism.
}}

\item{MWU}{A logical value that specifies the class of test statistic.
\itemize{
\item \code{MWU = FALSE} (first class statistic, default): rank-sum statistic.
\item \code{MWU = TRUE} (second class statistic): generalized Mann-Whitney U statistics.
}}

\item{method.list.treated}{A list that specifies the choice of the rank-based test statistic.
\itemize{
\item if \code{MWU = FALSE}, \code{method.list.treated} can be \code{list(name = "Wilcoxon")} or \code{list(name = "Stephenson", s = 10)}.
\item if \code{MWU = TRUE}, \code{method.list.treated} can be \code{list(name = "Wilcoxon")} or \code{list(name = "Polynomial", s = 10)}.
}}

\item{method.list.control}{A list that specifies the choice of the rank-based test statistic.
\itemize{
\item if \code{MWU = FALSE}, \code{method.list.control} can be \code{list(name = "Wilcoxon")} or \code{list(name = "Stephenson", s = 10)}.
\item if \code{MWU = TRUE}, \code{method.list.control} can be \code{list(name = "Wilcoxon")} or \code{list(name = "Polynomial", s = 10)}.
}}

\item{stat.null.treated}{An \eqn{nperm \times 1} vector whose empirical distribution
approximates the randomization distribution of the rank sum statistic.
\itemize{
\item if \code{stat.null.treated = NULL} (default), the function will calculate it internally.
\item if \code{missing = "general"}, \code{mp}, \code{mn}, then n choose n1.
\item if \code{missing = "sharp"}, \code{random}, then n.obs choose n1.obs.
}}

\item{Z.perm.treated}{A matrix that specifies the permutated assignments
for approximating the null distribution of the test statistic.
\itemize{
\item if \code{Z.perm.treated = NULL} (default), the function will calculate it internally.
\item if \code{missing = "general"}, \code{mp}, \code{mn}, then \code{Z.perm.treated} is an \eqn{n \times nperm} matrix (n choose n1).
\item if \code{missing = "sharp"}, \code{random}, then \code{Z.perm.treated} is an \eqn{n.obs \times nperm} matrix (n.obs choose n1.obs).
}}

\item{stat.null.control}{An \eqn{nperm \times 1} vector whose empirical distribution
approximates the randomization distribution of the rank sum statistic (n.obs choose n0.obs).
\itemize{
\item if \code{stat.null.control = NULL} (default), the function will calculate it internally.
\item if \code{missing = "general"}, \code{mp}, \code{mn}, then n choose n0.
\item if \code{missing = "sharp"}, \code{random}, then n.obs choose n0.obs.
}}

\item{Z.perm.control}{A matrix that specifies the permutated assignments
for approximating the null distribution of the test statistic.
\itemize{
\item if \code{Z.perm.control = NULL} (default), the function will calculate it internally.
\item if \code{missing = "general"}, \code{mp}, \code{mn}, then \code{Z.perm.control} is an \eqn{n \times nperm} matrix (n choose n0).
\item if \code{missing = "sharp"}, \code{random}, then \code{Z.perm.control} is an \eqn{n.obs \times nperm} matrix (n.obs choose n0.obs).
}}

\item{nperm}{A positive integer representing the number of permutations to
approximate the randomization distribution of the test statistic.}

\item{alpha}{A scalar, where \eqn{1-\alpha} indicates the confidence level.}

\item{tol}{A scalar that specifies the precision of the obtained confidence intervals.
For example, if \code{tol = 10^(-3)}, then the confidence limits are precise up to 3 digits.}
}
\value{
A dataframe with \eqn{L} rows and 3 columns (\eqn{k, lower, upper}).
Each row corresponds to the lower and upper confidence limits of the given \eqn{\tau_{(k)}}.
\itemize{
\item if \code{k.vec = NULL}, \eqn{L = n}.
\item if \code{k.vec} is not \code{NULL}, \eqn{L} is the length of \code{k.vec}.
}
}
\description{
\code{ci_quantile()} obtains one-sided confidence intervals
for all quantiles of individual treatment effects.
}
\examples{
Z <- c(1, 1, 1, 0, 0, 0)
Y <- c(8, 9, NA, 3, 4, NA)
conf.int <- ci_quantile(
  Z, Y, k.vec = NULL, missing = "general", MWU = FALSE,
  method.list.treated = list(name = "Wilcoxon"),
  method.list.control = list(name = "Wilcoxon"),
  stat.null.treated = NULL, Z.perm.treated = NULL,
  stat.null.control = NULL, Z.perm.control = NULL,
  nperm = 10^4, alpha = 0.05, tol = 10^(-3)
)
}
