% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_function_sharp_null.R
\name{ci_sharp_greater}
\alias{ci_sharp_greater}
\title{Confidence interval assuming constant treatment effect}
\usage{
ci_sharp_greater(
  Z,
  Y,
  missing = "general",
  MWU = FALSE,
  method.list = list(name = "Wilcoxon"),
  stat.null = NULL,
  Z.perm = NULL,
  nperm = 10^4,
  alpha = 0.05,
  tol = 10^(-3)
)
}
\arguments{
\item{Z}{Treatment assignment (\eqn{n \times 1} vector).}

\item{Y}{Observed outcome (\eqn{n \times 1} vector, including \code{NA}s).}

\item{missing}{A string specifying the missing mechanism.
\itemize{
\item \code{missing = "general"}: general missing mechanism.
\item \code{missing = "mp"}: monotone positive missing mechanism (\eqn{M_1 >= M_0}).
\item \code{missing = "mn"}: monotone negative missing mechanism (\eqn{M_1 <= M_0}).
\item \code{missing = "sharp"}: sharp missing mechanism (\eqn{M_1 = M_0}).
\item \code{missing = "random"}: missing at random mechanism.
}}

\item{MWU}{A logical value that specifies the class of test statistic.
\itemize{
\item \code{MWU = FALSE} (first class statistic, default): rank-sum statistic.
\item \code{MWU = TRUE} (second class statistic): generalized Mann-Whitney U statistic.
}}

\item{method.list}{A list that specifies the choice of the rank-based test statistic.
\itemize{
\item if \code{MWU = FALSE}, \code{method.list} can be \code{list(name = "Wilcoxon")} or \code{list(name = "Stephenson", s = 10)}.
\item if \code{MWU = TRUE}, \code{method.list} can be \code{list(name = "Wilcoxon")} or \code{list(name = "Polynomial", s = 10)}.
}}

\item{stat.null}{An \eqn{nperm \times 1} vector whose empirical distribution
approximates the randomization distribution of the rank-sum statistic.
\itemize{
\item if \code{stat.null = NULL} (default), the function will calculate it internally.
\item if \code{missing = "general"}, \code{mp}, \code{mn}, then n choose n1.
\item if \code{missing = "sharp"}, \code{random}, then n.obs choose n1.obs.
}}

\item{Z.perm}{A matrix that specifies the permutated assignments
for approximating the null distribution of the test statistic.
\itemize{
\item if \code{Z.perm = NULL} (default), the function will calculate it internally.
\item if \code{missing = "general"}, \code{mp}, \code{mn}, then \code{Z.perm} is an \eqn{n \times nperm} matrix (n choose n1).
\item if \code{missing = "sharp"}, \code{random}, then \code{Z.perm} is an \eqn{n.obs \times nperm} matrix (n.obs choose n1.obs).
}}

\item{nperm}{A positive integer representing the number of permutations to
approximate the randomization distribution of the test statistic.}

\item{alpha}{A scalar, where \eqn{1-\alpha} indicates the confidence level.}

\item{tol}{A numerical object that specifies the precision of the obtained confidence intervals.
For example, if \code{tol = 10^(-3)}, then the confidence limits are precise up to 3 digits.}
}
\value{
The one-sided confidence interval for the maximum individual effect.
}
\description{
\code{ci_sharp_greater()} obtains one-sided confidence interval for the maximum individual effect.
}
