% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_function_sharp_null_twostep.R
\name{ci_sharp_twostep}
\alias{ci_sharp_twostep}
\title{Confidence interval assuming constant treatment effect}
\usage{
ci_sharp_twostep(
  Z,
  Y,
  alternative,
  missing = "general",
  method.list = list(name = "Wilcoxon"),
  stat.null = NULL,
  Z.perm = NULL,
  nperm = 10^4,
  alpha = 0.05,
  beta = 0.1 * 0.05,
  tol = 10^(-3)
)
}
\arguments{
\item{Z}{Treatment assignment (\eqn{n \times 1} vector).}

\item{Y}{Observed outcome (\eqn{n \times 1} vector, including \code{NA}s).}

\item{alternative}{A string that specifies the direction of the alternative hypothesis:
\itemize{
\item \code{alternative = "greater"}: lower bound.
\item \code{alternative = "less"}: upper bound.
\item \code{alternative = "two.sided"}: (lower bound, upper bound).
}}

\item{missing}{A string specifying the missing mechanism.
\itemize{
\item \code{missing = "general"}: general missing mechanism.
\item \code{missing = "mp"}: monotone positive missing mechanism (\eqn{M_1 >= M_0}).
\item \code{missing = "mn"}: monotone negative missing mechanism (\eqn{M_1 <= M_0}).
}}

\item{method.list}{A list that specifies the choice of test statistic:
\code{method.list} can be \code{list(name = "Wilcoxon")} or \code{list(name = "Polynomial", s = 10)}.}

\item{stat.null}{An \eqn{nperm \times 1} vector whose empirical distribution
approximates the randomization distribution of the rank-based statistic
(\eqn{n} choose \eqn{n_1}).
\itemize{
\item if \code{stat.null = NULL} (default), the function will calculate it internally.
}}

\item{Z.perm}{An \eqn{n \times nperm} matrix that specifies the permuted assignments
for approximating the null distribution of the test statistic
(\eqn{n} choose \eqn{n_1}).
\itemize{
\item if \code{Z.perm = NULL} (default), the function will calculate it internally.
}}

\item{nperm}{A positive integer that specifies the number of permutations to
approximate the randomization distribution of the test statistic.}

\item{alpha}{A scalar, where \eqn{1-\alpha} indicates the confidence level.}

\item{beta}{A real number that belongs to \eqn{[0, \alpha]}.}

\item{tol}{A numerical object that specifies the precision of the obtained confidence intervals.
For example, if \code{tol = 10^(-3)}, then the confidence limits are precise up to 3 digits.}
}
\value{
The one-sided or two-sided confidence interval for the constant individual effect.
}
\description{
\code{ci_sharp_twostep()} obtains one-sided or two-sided confidence interval for the constant individual effect.
}
