% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_function_quantile.R
\name{conf_quant_treated_g_mp_mn}
\alias{conf_quant_treated_g_mp_mn}
\title{Calculate one-sided confidence interval (c, Inf) for all quantiles among treated units with missing outcomes for general/mp/mn missing}
\usage{
conf_quant_treated_g_mp_mn(
  Z,
  Y,
  k.vec = NULL,
  missing = "general",
  MWU = FALSE,
  method.list = list(name = "Wilcoxon"),
  stat.null = NULL,
  Z.perm = NULL,
  nperm = 10^4,
  alpha = 0.05,
  tol = 10^(-3)
)
}
\arguments{
\item{Z}{Treatment assignment (\eqn{n \times 1} vector).}

\item{Y}{Observed outcome (\eqn{n \times 1} vector, including \code{NA}s).}

\item{k.vec}{A vector that specifies the quantiles of individual effects of interest.
If \code{k.vec = NULL}, then we consider all quantiles of individual effects \eqn{1 \leq k \leq n_1}.}

\item{missing}{A string specifying the missing mechanism.
\itemize{
\item \code{missing = "general"}: general missing mechanism.
\item \code{missing = "mp"}: monotone positive missing mechanism (\eqn{M_1 >= M_0}).
\item \code{missing = "mn"}: monotone negative missing mechanism (\eqn{M_1 <= M_0}).
}}

\item{MWU}{A logical value that specifies the class of test statistic.
\itemize{
\item \code{MWU = FALSE} (first class statistic, default): rank-sum statistic.
\item \code{MWU = TRUE} (second class statistic): generalized Mann-Whitney U statistic.
}}

\item{method.list}{A list that specifies the choice of the rank-based test statistic.
\itemize{
\item if \code{MWU = FALSE}, \code{method.list} can be \code{list(name = "Wilcoxon")} or \code{list(name = "Stephenson", s = 10)}.
\item if \code{MWU = TRUE}, \code{method.list} can be \code{list(name = "Wilcoxon")} or \code{list(name = "Polynomial", s = 10)}.
}}

\item{stat.null}{An \eqn{nperm \times 1} vector whose empirical distribution
approximates the randomization distribution of the rank sum statistic (n choose n1).
\itemize{
\item if \code{stat.null = NULL} (default), the function will calculate it internally.
}}

\item{Z.perm}{An \eqn{n \times nperm} matrix that specifies the permutated assignments
for approximating the null distribution of the test statistic (n choose n1).
\itemize{
\item if \code{Z.perm = NULL} (default), the function will calculate it internally.
}}

\item{nperm}{A positive integer representing the number of permutations to
approximate the randomization distribution of the test statistic.}

\item{alpha}{A scalar, where \eqn{1-\alpha} indicates the confidence level.}

\item{tol}{A scalar that specifies the precision of the obtained confidence intervals.
For example, if \code{tol = 10^(-3)}, then the confidence limits are precise up to 3 digits.}
}
\value{
An \eqn{L \times 1} vector corresponding to the lower confidence limits of the given \eqn{\tau_{(k)}}.
\itemize{
\item if \code{k.vec = NULL}, \eqn{L = n_{1}}.
\item if \code{k.vec} is not \code{NULL}, \eqn{L} is the length of \code{k.vec}.
}
}
\description{
Calculate one-sided confidence interval (c, Inf) for all quantiles among treated units with missing outcomes for general/mp/mn missing
}
\examples{
Z <- c(1, 1, 1, 0, 0, 0)
Y <- c(8, 9, NA, 3, 4, NA)
c.limit <- conf_quant_treated_g_mp_mn(Z, Y,
  k.vec = NULL, missing = "general",
  MWU = FALSE, method.list = list(name = "Wilcoxon"),
  stat.null = NULL, Z.perm = NULL, nperm = 10^4, alpha = 0.05, tol = 10^(-3)
)
}
