% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_function_sharp_null.R
\name{lci}
\alias{lci}
\title{Step 1 of two-step method: construct confidence intervals of hypergeometric parameter}
\usage{
lci(N, x, n, alpha)
}
\arguments{
\item{N}{Number of total units.}

\item{x}{Number of observed good units.}

\item{n}{Number of draws.}

\item{alpha}{Confidence parameter.}
}
\value{
The lower limit of \eqn{1 - \alpha} CP interval.
}
\description{
Generate the lower limit of \eqn{1 - \alpha} CP interval.
Consider hypergeometric with \eqn{N} units, \eqn{M} of them are good (\eqn{N - M} are bad),
select \eqn{n} from \eqn{N} units, there are \eqn{x} good units in the sample.
P(X = x) = dhyper(x, m, N - m, n) for max(0, n + M - N) <= x <= min(n, M)
Here, \eqn{n} and \eqn{N} are known, and we want to estimate \eqn{M} based on \eqn{x}.
Source: "Exact Optimal Confidence Intervals for Hypergeometric Parameters"
https://doi.org/10.1080/01621459.2014.966191
}
\examples{
lci(N = 100, x = 50, n = 50, alpha = 0.1)
}
