% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lee2009}
\alias{lee2009}
\title{National Job Corps Study data}
\format{
A data frame with 9145 rows and 210 columns.
}
\source{
Replication data: \url{https://www.openicpsr.org/openicpsr/project/113269/version/V1/view?flag=follow&pageSize=100&sortOrder=(?title)&sortAsc=true}
Replication code: \url{https://academic.oup.com/restud/article/76/3/1071/1590707?guestAccessKey=#supplementary-data}
}
\usage{
lee2009
}
\description{
Replication data for Lee (2009)
}
\keyword{datasets}
