% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_function_quantile.R
\name{min_stat}
\alias{min_stat}
\title{Calculate minimum test statistic value in \eqn{H_{k,c}} with no missing outcomes}
\usage{
min_stat(Z, Y, k, c, MWU = FALSE, method.list = list(name = "Wilcoxon"))
}
\arguments{
\item{Z}{Treatment assignment (\eqn{n \times 1} vector).}

\item{Y}{Observed outcome (\eqn{n \times 1} vector, no \code{NA}s).}

\item{k}{A positive integer that specifies which quantile of individual effect is of interest (\eqn{1 \leq k \leq n_1}).}

\item{c}{A scalar or vector specifying the sharp null hypothesis.}

\item{MWU}{A logical value that specifies the class of test statistic.
\itemize{
\item \code{MWU = FALSE} (first class statistic, default): rank-sum statistic.
\item \code{MWU = TRUE} (second class statistic): generalized Mann-Whitney U statistic.
}}

\item{method.list}{A list that specifies the choice of the rank-based test statistic.
\itemize{
\item if \code{MWU = FALSE}, \code{method.list} can be \code{list(name = "Wilcoxon")} or \code{list(name = "Stephenson", s = 10)}.
\item if \code{MWU = TRUE}, \code{method.list} can be \code{list(name = "Wilcoxon")} or \code{list(name = "Polynomial", s = 10)}.
}}
}
\value{
A scalar.
}
\description{
Calculate minimum test statistic value in \eqn{H_{k,c}} with no missing outcomes
}
\examples{
Z <- c(1, 1, 1, 0, 0, 0)
Y <- c(7, 8, 6, 3, 4, 1)
min_stat(Z, Y, k = 2, c = 0, MWU = FALSE, method.list = list(name = "Wilcoxon"))
}
