% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_function_quantile.R
\name{min_stat_g_mp_mn}
\alias{min_stat_g_mp_mn}
\title{Calculate minimum test statistic value in \eqn{H_{k,c}} with missing outcomes for general/mp/mn missing}
\usage{
min_stat_g_mp_mn(
  Z,
  Y,
  k,
  c,
  missing = "general",
  MWU = FALSE,
  method.list = list(name = "Wilcoxon")
)
}
\arguments{
\item{Z}{Treatment assignment (\eqn{n \times 1} vector).}

\item{Y}{Observed outcome (\eqn{n \times 1} vector, including \code{NA}s).}

\item{k}{A positive integer that specifies which quantile of individual effect is of interest (\eqn{1 \leq k \leq n_1}).}

\item{c}{A scalar or vector specifying the sharp null hypothesis.}

\item{missing}{A string specifying the missing mechanism.
\itemize{
\item \code{missing = "general"}: general missing mechanism.
\item \code{missing = "mp"}: monotone positive missing mechanism (\eqn{M_1 >= M_0}).
\item \code{missing = "mn"}: monotone negative missing mechanism (\eqn{M_1 <= M_0}).
}}

\item{MWU}{A logical value that specifies the class of test statistic.
\itemize{
\item \code{MWU = FALSE} (first class statistic, default): rank-sum statistic.
\item \code{MWU = TRUE} (second class statistic): generalized Mann-Whitney U statistic.
}}

\item{method.list}{A list that specifies the choice of the rank-based test statistic.
\itemize{
\item if \code{MWU = FALSE}, \code{method.list} can be \code{list(name = "Wilcoxon")} or \code{list(name = "Stephenson", s = 10)}.
\item if \code{MWU = TRUE}, \code{method.list} can be \code{list(name = "Wilcoxon")} or \code{list(name = "Polynomial", s = 10)}.
}}
}
\value{
A scalar.
}
\description{
Calculate minimum test statistic value in \eqn{H_{k,c}} with missing outcomes for general/mp/mn missing
}
\examples{
Z <- c(1, 1, 1, 0, 0, 0)
Y <- c(NA, 8, 6, 3, 4, NA)
min_stat_g_mp_mn(
  Z, Y, k = 2, c = 0, missing = "general",
  MWU = FALSE, method.list = list(name = "Wilcoxon")
)
}
