% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_function_sharp_null.R
\name{null_dist}
\alias{null_dist}
\title{Generate randomization distribution of the rank-based test statistic}
\usage{
null_dist(
  n,
  m,
  class = "RS",
  method.list = list(name = "Wilcoxon"),
  Z.perm = NULL,
  nperm = 10^4
)
}
\arguments{
\item{n}{A positive integer that specifies the number of units.}

\item{m}{A positive integer that specifies the number of treated units.}

\item{class}{A string that specifies the class of test statistic:
\itemize{
\item \code{class = "RS"} (first class, default): rank-sum statistic.
\item \code{class = "MWU+"} (second class): generalized Mann-Whitney U statistic,
defined as the sum of relative ranks for treated units.
\item \code{class = "MWU-"} (second class): generalized Mann-Whitney U statistic,
defined as the negative sum of relative ranks for control units.
}}

\item{method.list}{A list that specifies the choice of test statistic:
\itemize{
\item if \code{class = "RS"}, \code{method.list} can be \code{list(name = "Wilcoxon")} or
\code{list(name = "Stephenson", s = 10)}.
\item if \code{class = "MWU+"} or \code{class = "MWU-"}, \code{method.list} can be
\code{list(name = "Wilcoxon")} or \code{list(name = "Polynomial", s = 10)}.
}}

\item{Z.perm}{An \eqn{n \times nperm} matrix that specifies the permuted assignments
for approximating the null distribution of the test statistic
(\eqn{n} choose \eqn{m}).
\itemize{
\item if \code{Z.perm = NULL} (default), the function will calculate it internally.
}}

\item{nperm}{A positive integer that specifies the number of permutations to
approximate the randomization distribution of the test statistic.}
}
\value{
An \eqn{nperm \times 1} vector.
}
\description{
\code{null_dist()} generates the null distribution of the given rank-based test
statistic for an experiment with \code{m} treated out of \code{n} units.
}
\examples{
null_dist(
  n = 5, m = 3, class = "RS", method.list = list(name = "Wilcoxon"),
  Z.perm = NULL, nperm = 10^4
)
null_dist(
  n = 5, m = 3, class = "MWU+", method.list = list(name = "Wilcoxon"),
  Z.perm = NULL, nperm = 10^4
)
null_dist(
  n = 5, m = 3, class = "MWU-", method.list = list(name = "Wilcoxon"),
  Z.perm = NULL, nperm = 10^4
)
}
