% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_function_sharp_null_twostep.R
\name{psi}
\alias{psi}
\title{Indicator function for pairwise comparison}
\usage{
psi(i, j, x, y)
}
\arguments{
\item{i}{Index for x.}

\item{j}{Index for y.}

\item{x}{A scalar.}

\item{y}{A scalar.}
}
\value{
The indicator function for pairwise comparison.
}
\description{
Indicator function for pairwise comparison
}
