% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_function_quantile.R
\name{pval_quantile}
\alias{pval_quantile}
\title{Randomization test for quantiles of individual treatment effects with missing outcomes}
\usage{
pval_quantile(
  Z,
  Y,
  k,
  c,
  missing = "general",
  MWU = FALSE,
  method.list = list(name = "Wilcoxon"),
  stat.null = NULL,
  Z.perm = NULL,
  nperm = 10^4
)
}
\arguments{
\item{Z}{Treatment assignment (\eqn{n \times 1} vector).}

\item{Y}{Observed outcome (\eqn{n \times 1} vector, including \code{NA}s).}

\item{k}{A positive integer that specifies which quantile of individual effect is of interest
\itemize{
\item if \code{missing = "general"} or \code{missing = "mp"} or \code{missing = "mn"}, \eqn{1 \leq k \leq n_1}.
\item if \code{missing = "sharp"} or \code{missing = "random"}, \eqn{1 \leq k \leq n_{11}}.
}}

\item{c}{A scalar or vector specifying the sharp null hypothesis.}

\item{missing}{A string specifying the missing mechanism.
\itemize{
\item \code{missing = "general"}: general missing mechanism.
\item \code{missing = "mp"}: monotone positive missing mechanism (\eqn{M_1 >= M_0}).
\item \code{missing = "mn"}: monotone negative missing mechanism (\eqn{M_1 <= M_0}).
\item \code{missing = "sharp"}: sharp missing mechanism (\eqn{M_1 = M_0}).
\item \code{missing = "random"}: random missing mechanism.
}}

\item{MWU}{A logical value that specifies the class of test statistic.
\itemize{
\item \code{MWU = FALSE} (first class statistic, default): rank-sum statistic.
\item \code{MWU = TRUE} (second class statistic): generalized Mann-Whitney U statistics.
}}

\item{method.list}{A list that specifies the choice of the rank-based test statistic.
\itemize{
\item if \code{MWU = FALSE}, \code{method.list} can be \code{list(name = "Wilcoxon")} or \code{list(name = "Stephenson", s = 10)}.
\item if \code{MWU = TRUE}, \code{method.list} can be \code{list(name = "Wilcoxon")} or \code{list(name = "Polynomial", s = 10)}.
}}

\item{stat.null}{An \eqn{nperm \times 1} vector whose empirical distribution
approximates the randomization distribution of the rank sum statistic.
\itemize{
\item if \code{stat.null = NULL} (default), the function will calculate it internally.
\item if \code{missing = "general"}, \code{mp}, \code{mn}, then n choose n1.
\item if \code{missing = "sharp"}, \code{random}, then n.obs choose n1.obs.
}}

\item{Z.perm}{A matrix that specifies the permutated assignments
for approximating the null distribution of the test statistic.
\itemize{
\item if \code{Z.perm = NULL} (default), the function will calculate it internally.
\item if \code{missing = "general"}, \code{mp}, \code{mn}, then \code{Z.perm} is an \eqn{n \times nperm} matrix (n choose n1).
\item if \code{missing = "sharp"}, \code{random}, then \code{Z.perm} is an \eqn{n.obs \times nperm} matrix (n.obs choose n1.obs).
}}

\item{nperm}{A positive integer representing the number of permutations to
approximate the randomization distribution of the test statistic.}
}
\value{
The p-value for testing the specified null hypothesis of interest (a scalar).
}
\description{
\code{pval_quantile()} obtains the p-value for testing the null hypothesis
\eqn{H_{k,c}: \tau_{(k)} \leq c}, where \eqn{\tau_{(k)}} denotes individual treatment effect at rank \eqn{k}.
}
