% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_function_quantile.R
\name{pval_quantile_s_r}
\alias{pval_quantile_s_r}
\title{Randomization test for quantiles of individual treatment effects with missing outcomes for sharp/random missing}
\usage{
pval_quantile_s_r(
  Z,
  Y,
  k,
  c,
  MWU = FALSE,
  method.list = list(name = "Wilcoxon"),
  stat.null = NULL,
  Z.perm = NULL,
  nperm = 10^4
)
}
\arguments{
\item{Z}{Treatment assignment (\eqn{n \times 1} vector).}

\item{Y}{Observed outcome (\eqn{n \times 1} vector, including \code{NA}s).}

\item{k}{A positive integer that specifies which quantile of individual effect is of interest (\eqn{1 \leq k \leq n_{11}})}

\item{c}{A scalar or vector specifying the sharp null hypothesis.}

\item{MWU}{A logical value that specifies the class of test statistic.
\itemize{
\item \code{MWU = FALSE} (first class statistic, default): rank-sum statistic.
\item \code{MWU = TRUE} (second class statistic): generalized Mann-Whitney U statistic.
}}

\item{method.list}{A list that specifies the choice of the rank-based test statistic.
\itemize{
\item if \code{MWU = FALSE}, \code{method.list} can be \code{list(name = "Wilcoxon")} or \code{list(name = "Stephenson", s = 10)}.
\item if \code{MWU = TRUE}, \code{method.list} can be \code{list(name = "Wilcoxon")} or \code{list(name = "Polynomial", s = 10)}.
}}

\item{stat.null}{An \eqn{nperm \times 1} vector whose empirical distribution
approximates the randomization distribution of the rank sum statistic (n.obs choose n1.obs).
\itemize{
\item if \code{stat.null = NULL} (default), the function will calculate it internally.
}}

\item{Z.perm}{An \eqn{n.obs \times nperm} matrix that specifies the permutated assignments
for approximating the null distribution of the test statistic (n.obs choose n1.obs).
\itemize{
\item if \code{Z.perm = NULL} (default), the function will calculate it internally.
}}

\item{nperm}{A positive integer representing the number of permutations to
approximate the randomization distribution of the test statistic.}
}
\value{
The p-value for testing the specified null hypothesis of interest (a scalar).
}
\description{
\code{pval_quantile_s_r()} obtains the p-value for testing the null hypothesis
\eqn{H_{k,c}: \tau_{(k)} \leq c}, where \eqn{\tau_{(k)}} denotes individual treatment effect at rank \eqn{k}.
Naive approach: drop missing data and set \eqn{n_{11}=\sum_{i=1}^n Z_i*M_i} treated
and \eqn{n_{01}=\sum_{i=1}^n (1-Z_i)*M_i} control.
Implement CRE for only units without missing outcomes,
and perform usual randomization test with \eqn{n_{11}} treated and \eqn{n_{01}} control.
}
\examples{
Z <- c(1, 1, 1, 0, 0, 0)
Y <- c(8, 9, NA, 3, 4, NA)
pval <- pval_quantile_s_r(
  Z, Y, k = 2, c = 0, MWU = FALSE, method.list = list(name = "Wilcoxon"),
  stat.null = NULL, Z.perm = NULL, nperm = 10^4
)
}
