% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_function_sharp_null.R
\name{pval_sharp_composite}
\alias{pval_sharp_composite}
\title{Randomization test for sharp null hypotheses with missing outcomes}
\usage{
pval_sharp_composite(
  Z,
  Y,
  c = 0,
  b00,
  b01,
  b10,
  missing = "general",
  class = "RS",
  method.list = list(name = "Wilcoxon"),
  stat.null = NULL,
  Z.perm = NULL,
  nperm = 10^4
)
}
\arguments{
\item{Z}{Treatment assignment (\eqn{n \times 1} vector).}

\item{Y}{Observed outcome (\eqn{n \times 1} vector, including \code{NA}s).}

\item{c}{A scalar that specifies the sharp null hypothesis.}

\item{b00, b01, b10}{Three scalars that specifies the composite control potential outcome.}

\item{missing}{A string that specifies the missing mechanism:
\itemize{
\item \code{missing = "general"}: general missing mechanism.
\item \code{missing = "mp"}: monotone positive missing mechanism (\eqn{M_1 \geq M_0}).
\item \code{missing = "mn"}: monotone negative missing mechanism (\eqn{M_1 \leq M_0}).
\item \code{missing = "sharp"}: sharp missing mechanism (\eqn{M_1 = M_0}).
}}

\item{class}{A string that specifies the class of test statistic:
\itemize{
\item \code{class = "RS"} (first class, default): rank-sum statistic.
\item \code{class = "MWU+"} (second class): generalized Mann-Whitney U statistic,
defined as the sum of relative ranks for treated units.
\item \code{class = "MWU-"} (second class): generalized Mann-Whitney U statistic,
defined as the negative sum of relative ranks for control units.
}}

\item{method.list}{A list that specifies the choice of test statistic:
\itemize{
\item if \code{class = "RS"}, \code{method.list} can be \code{list(name = "Wilcoxon")} or
\code{list(name = "Stephenson", s = 10)}.
\item if \code{class = "MWU+"} or \code{class = "MWU-"}, \code{method.list} can be
\code{list(name = "Wilcoxon")} or \code{list(name = "Polynomial", s = 10)}.
}}

\item{stat.null}{An \eqn{nperm \times 1} vector whose empirical distribution
approximates the randomization distribution of the rank-based statistic
(\eqn{n} choose \eqn{n_1}).
\itemize{
\item if \code{stat.null = NULL} (default), the function will calculate it internally.
}}

\item{Z.perm}{An \eqn{n \times nperm} matrix that specifies the permuted assignments
for approximating the null distribution of the test statistic
(\eqn{n} choose \eqn{n_1}).
\itemize{
\item if \code{Z.perm = NULL} (default), the function will calculate it internally.
}}

\item{nperm}{A positive integer that specifies the number of permutations to
approximate the randomization distribution of the test statistic.}
}
\value{
The p-value for testing the specified sharp null hypothesis of interest.
}
\description{
\code{pval_sharp_composite()} obtains the p-value for testing the sharp null hypothesis
\eqn{H_0: \tau = c}, using the worst-case imputed composite control potential outcome \eqn{Y_b(0)}.
}
\examples{
Z <- c(1, 1, 1, 0, 0, 0)
Y <- c(NA, 8, 6, 3, 4, NA)
pval_sharp_composite(Z, Y,
  c = 0, b00 = 0, b01 = Inf, b10 = -Inf, missing = "general",
  class = "RS", method.list = list(name = "Wilcoxon"),
  stat.null = NULL, Z.perm = NULL, nperm = 10^4
)
}
