% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_function_sharp_null.R
\name{pval_sharp_naive}
\alias{pval_sharp_naive}
\title{Randomization test for sharp null hypotheses with missing outcomes}
\usage{
pval_sharp_naive(
  Z,
  Y,
  c = 0,
  class = "RS",
  method.list = list(name = "Wilcoxon"),
  stat.null = NULL,
  Z.perm = NULL,
  nperm = 10^4
)
}
\arguments{
\item{Z}{Treatment assignment (\eqn{n \times 1} vector).}

\item{Y}{Observed outcome (\eqn{n \times 1} vector, including \code{NA}s).}

\item{c}{A scalar that specifies the sharp null hypothesis.}

\item{class}{A string that specifies the class of test statistic:
\itemize{
\item \code{class = "RS"} (first class, default): rank-sum statistic.
\item \code{class = "MWU+"} (second class): generalized Mann-Whitney U statistic,
defined as the sum of relative ranks for treated units.
\item \code{class = "MWU-"} (second class): generalized Mann-Whitney U statistic,
defined as the negative sum of relative ranks for control units.
}}

\item{method.list}{A list that specifies the choice of test statistic:
\itemize{
\item if \code{class = "RS"}, \code{method.list} can be \code{list(name = "Wilcoxon")} or
\code{list(name = "Stephenson", s = 10)}.
\item if \code{class = "MWU+"} or \code{class = "MWU-"}, \code{method.list} can be
\code{list(name = "Wilcoxon")} or \code{list(name = "Polynomial", s = 10)}.
}}

\item{stat.null}{An \eqn{nperm \times 1} vector whose empirical distribution
approximates the randomization distribution of the rank-based statistic
(\eqn{n_{11} + n_{01}} choose \eqn{n_{11}}).
\itemize{
\item if \code{stat.null = NULL} (default), the function will calculate it internally.
}}

\item{Z.perm}{An \eqn{(n_{11} + n_{01}) \times nperm} matrix that specifies the permuted assignments
for approximating the null distribution of the test statistic
(\eqn{n_{11} + n_{01}} choose \eqn{n_{11}}).
\itemize{
\item if \code{Z.perm = NULL} (default), the function will calculate it internally.
}}

\item{nperm}{A positive integer that specifies the number of permutations to
approximate the randomization distribution of the test statistic.}
}
\value{
The p-value for testing the specified sharp null hypothesis of interest.
}
\description{
\code{pval_sharp_naive()} obtains the p-value for testing the sharp null hypothesis
\eqn{H_0: \tau = c} using naive approach. Implement CRE for only units without
missing outcomes, and perform usual randomization test with \eqn{n_{11}} treated
and \eqn{n_{01}} control.
}
\examples{
Z <- c(1, 1, 1, 0, 0, 0)
Y <- c(NA, 8, 6, 3, 4, NA)
pval_sharp_naive(Z, Y,
  c = 0, class = "RS", method.list = list(name = "Wilcoxon"),
  stat.null = NULL, Z.perm = NULL, nperm = 10^4
)
}
