% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_function_sharp_null.R
\name{rank_control_relative_to_treat}
\alias{rank_control_relative_to_treat}
\title{Calculate the rank of control units relative to treated units}
\usage{
rank_control_relative_to_treat(Z, Y)
}
\arguments{
\item{Z}{Treatment assignment (\eqn{n \times 1} vector).}

\item{Y}{Realized outcome (\eqn{n \times 1} vector, no \code{NA}s).}
}
\value{
A \eqn{n_0 \times 1} vector
}
\description{
Calculate the rank of control units relative to treated units
}
\examples{
Z <- c(1, 0, 0, 1)
Y <- c(1:4)
rank_control_relative_to_treat(Z, Y)
# c(1, 1)
}
