% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_function_quantile.R
\name{threshold_correct_hypergeom}
\alias{threshold_correct_hypergeom}
\title{Calculate the correction term Delta using the proposed choice of kk.vec (k_prime)}
\usage{
threshold_correct_hypergeom(
  N = 100,
  n = 50,
  K.vec = c(60, 80),
  alpha = 0.05,
  ndraw = 10^4,
  hg.draw = NULL,
  tol = 10^(-2)
)
}
\description{
Calculate the correction term Delta using the proposed choice of kk.vec (k_prime)
}
