% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_function_sharp_null.R
\name{assign_CRE}
\alias{assign_CRE}
\title{Generate treatment assignments from a completely randomized experiment}
\usage{
assign_CRE(n, m, nperm)
}
\arguments{
\item{n}{A positive integer that specifies the number of units.}

\item{m}{A positive integer that specifies the number of treated units.}

\item{nperm}{A positive integer that specifies the number of permutations to
approximate the randomization distribution of the test statistic.}
}
\value{
An \eqn{n \times nperm} matrix.
}
\description{
\code{assign_CRE()} draws \code{nperm} independent assignments, each assigning
exactly \code{m} of \code{n} units to treatment.
}
\examples{
assign_CRE(n = 5, m = 3, nperm = 10^4)
}
