% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_function_sharp_null.R
\name{ci_sharp}
\alias{ci_sharp}
\title{Confidence interval assuming constant treatment effect}
\usage{
ci_sharp(
  Z,
  Y,
  alternative,
  missing = "general",
  class = "RS",
  method.list = list(name = "Wilcoxon"),
  stat.null = NULL,
  Z.perm = NULL,
  nperm = 10^4,
  alpha = 0.05,
  tol = 10^(-3)
)
}
\arguments{
\item{Z}{Treatment assignment (\eqn{n \times 1} vector).}

\item{Y}{Observed outcome (\eqn{n \times 1} vector, including \code{NA}s).}

\item{alternative}{A string that specifies the direction of the alternative hypothesis:
\itemize{
\item \code{alternative = "greater"}: lower bound.
\item \code{alternative = "less"}: upper bound.
\item \code{alternative = "two.sided"}: (lower bound, upper bound).
}}

\item{missing}{A string that specifies the missing mechanism:
\itemize{
\item \code{missing = "general"}: general missing mechanism.
\item \code{missing = "mp"}: monotone positive missing mechanism (\eqn{M_1 \geq M_0}).
\item \code{missing = "mn"}: monotone negative missing mechanism (\eqn{M_1 \leq M_0}).
\item \code{missing = "sharp"}: sharp missing mechanism (\eqn{M_1 = M_0}).
\item \code{missing = "random"}: missing at random mechanism.
}}

\item{class}{A string that specifies the class of test statistic:
\itemize{
\item \code{class = "RS"} (first class, default): rank-sum statistic.
\item \code{class = "MWU+"} (second class): generalized Mann-Whitney U statistic,
defined as the sum of relative ranks for treated units.
\item \code{class = "MWU-"} (second class): generalized Mann-Whitney U statistic,
defined as the negative sum of relative ranks for control units.
}}

\item{method.list}{A list that specifies the choice of test statistic:
\itemize{
\item if \code{class = "RS"}, \code{method.list} can be \code{list(name = "Wilcoxon")} or
\code{list(name = "Stephenson", s = 10)}.
\item if \code{class = "MWU+"} or \code{class = "MWU-"}, \code{method.list} can be
\code{list(name = "Wilcoxon")} or \code{list(name = "Polynomial", s = 10)}.
}}

\item{stat.null}{An \eqn{nperm \times 1} vector whose empirical distribution
approximates the randomization distribution of the rank-based statistic.
\itemize{
\item if \code{stat.null = NULL} (default), the function will calculate it internally.
\item if \code{missing = "general"}, \code{mp}, \code{mn}, then \eqn{n} choose \eqn{n_1}.
\item if \code{missing = "sharp"}, \code{random}, then \eqn{n_{11} + n_{01}} choose \eqn{n_{11}}.
}}

\item{Z.perm}{A matrix that specifies the permuted assignments
for approximating the null distribution of the test statistic.
\itemize{
\item if \code{Z.perm = NULL} (default), the function will calculate it internally.
\item if \code{missing = "general"}, \code{mp}, \code{mn}, then \code{Z.perm} is an \eqn{n \times nperm} matrix
(\eqn{n} choose \eqn{n_1}).
\item if \code{missing = "sharp"}, \code{random}, then \code{Z.perm} is an \eqn{(n_{11} + n_{01}) \times nperm} matrix
(\eqn{n_{11} + n_{01}} choose \eqn{n_{11}}).
}}

\item{nperm}{A positive integer that specifies the number of permutations to
approximate the randomization distribution of the test statistic.}

\item{alpha}{A scalar, where \eqn{1-\alpha} indicates the confidence level.}

\item{tol}{A numerical object that specifies the precision of the obtained confidence intervals.
For example, if \code{tol = 10^(-3)}, then the confidence limits are precise up to 3 digits.}
}
\value{
The one-sided or two-sided confidence interval for the maximum individual effect.
}
\description{
\code{ci_sharp()} obtains one-sided or two-sided confidence interval for the maximum individual effect.
}
