% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_function_sharp_null.R
\name{null_dist}
\alias{null_dist}
\title{Generate randomization distribution of the rank-based test statistic}
\usage{
null_dist(
  n,
  m,
  MWU = FALSE,
  method.list = list(name = "Wilcoxon"),
  Z.perm = NULL,
  nperm = 10^4
)
}
\arguments{
\item{n}{A positive integer that specifies the number of units.}

\item{m}{A positive integer that specifies the number of treated units.}

\item{MWU}{A logical value that specifies the class of test statistic.
\itemize{
\item \code{MWU = FALSE} (first class statistic, default): rank-sum statistic.
\item \code{MWU = TRUE} (second class statistic): generalized Mann-Whitney U statistic.
}}

\item{method.list}{A list that specifies the choice of the rank-based test statistic.
\itemize{
\item if \code{MWU = FALSE}, \code{method.list} can be \code{list(name = "Wilcoxon")} or \code{list(name = "Stephenson", s = 10)}.
\item if \code{MWU = TRUE}, \code{method.list} can be \code{list(name = "Wilcoxon")} or \code{list(name = "Polynomial", s = 10)}.
}}

\item{Z.perm}{An \eqn{n \times nperm} matrix that specifies the permutated assignments
for approximating the null distribution of the test statistic (n choose m).
\itemize{
\item if \code{Z.perm = NULL} (default), the function will calculate it internally.
}}

\item{nperm}{A positive integer representing the number of permutations to
approximate the randomization distribution of the test statistic.}
}
\value{
An \eqn{nperm \times 1} vector.
}
\description{
Generate the null distribution of the given rank-based test statistic
for an experiment with m treated out of n units.
}
\examples{
null_dist(
  n = 5, m = 3, MWU = FALSE, method.list = list(name = "Wilcoxon"),
  Z.perm = NULL, nperm = 10^4
)
null_dist(
  n = 5, m = 3, MWU = TRUE, method.list = list(name = "Wilcoxon"),
  Z.perm = NULL, nperm = 10^4
)
}
