% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_function_sharp_null_twostep.R
\name{pval_sharp_mp_twostep}
\alias{pval_sharp_mp_twostep}
\title{Randomization test for sharp null hypotheses with missing outcomes}
\usage{
pval_sharp_mp_twostep(
  Z,
  Y,
  c = 0,
  method.list = list(name = "Wilcoxon"),
  stat.null = NULL,
  Z.perm = NULL,
  nperm = 10^4,
  beta
)
}
\arguments{
\item{Z}{Treatment assignment (\eqn{n \times 1} vector).}

\item{Y}{Observed outcome (\eqn{n \times 1} vector, including \code{NA}s).}

\item{c}{A scalar that specifies the sharp null hypothesis.}

\item{method.list}{A list that specifies the choice of test statistic:
\code{method.list} can be \code{list(name = "Wilcoxon")} or \code{list(name = "Polynomial", s = 10)}.}

\item{stat.null}{An \eqn{nperm \times 1} vector whose empirical distribution
approximates the randomization distribution of the rank-based statistic
(\eqn{n} choose \eqn{n_1}).
\itemize{
\item if \code{stat.null = NULL} (default), the function will calculate it internally.
}}

\item{Z.perm}{An \eqn{n \times nperm} matrix that specifies the permuted assignments
for approximating the null distribution of the test statistic
(\eqn{n} choose \eqn{n_1}).
\itemize{
\item if \code{Z.perm = NULL} (default), the function will calculate it internally.
}}

\item{nperm}{A positive integer that specifies the number of permutations to
approximate the randomization distribution of the test statistic.}

\item{beta}{A real number that belongs to \eqn{[0, \alpha]}.}
}
\value{
The p-value for testing the specified sharp null hypothesis of interest.
}
\description{
\code{pval_sharp_mp_twostep()} obtains the p-value for testing the sharp null hypothesis
\eqn{H_0: \tau = c}, using the two-step procedure under monotone positive missing mechanism.
}
