% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_function_sharp_null.R
\name{pval_sharp_twostep}
\alias{pval_sharp_twostep}
\title{Randomization test for sharp null hypotheses with missing outcomes}
\usage{
pval_sharp_twostep(
  Z,
  Y,
  c = 0,
  missing,
  method.list = list(name = "Wilcoxon"),
  stat.null = NULL,
  Z.perm = NULL,
  nperm = 10^4,
  beta
)
}
\arguments{
\item{Z}{Treatment assignment (\eqn{n \times 1} vector).}

\item{Y}{Observed outcome (\eqn{n \times 1} vector, including \code{NA}s).}

\item{c}{A scalar or vector specifying the sharp null hypothesis.}

\item{missing}{A string specifying the missing mechanism.
\itemize{
\item \code{missing = "general"}: general missing mechanism.
\item \code{missing = "mp"}: monotone positive missing mechanism (\eqn{M_1 >= M_0}).
\item \code{missing = "mn"}: monotone negative missing mechanism (\eqn{M_1 <= M_0}).
}}

\item{method.list}{A list that specifies the choice of the rank-based test statistic.
method.list\verb{can be}list(name = "Wilcoxon")\code{or}list(name = "Polynomial", s = 10)`.}

\item{stat.null}{An \eqn{nperm \times 1} vector whose empirical distribution
approximates the randomization distribution of the rank-sum statistic (n choose n1).
\itemize{
\item if \code{stat.null = NULL} (default), the function will calculate it internally.
}}

\item{Z.perm}{An \eqn{n \times nperm} matrix that specifies the permutated assignments
for approximating the null distribution of the test statistic (n choose n1).
\itemize{
\item if \code{Z.perm = NULL} (default), the function will calculate it internally.
}}

\item{nperm}{A positive integer representing the number of permutations to
approximate the randomization distribution of the test statistic.}

\item{beta}{A real number that belongs to \eqn{[0, \alpha]}.}
}
\value{
The p-value for testing the specified sharp null hypothesis of interest.
}
\description{
\code{pval_sharp_twostep()} obtains the p-value for testing the sharp null hypothesis \eqn{H_0: \tau = c}.

Approach 3: use the two-step procedure and the worst-case imputed composite control potential outcome
to calculate the test statistic (b = Inf).
}
