% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_function_sharp_null.R
\name{rank_score}
\alias{rank_score}
\title{Calculate the rank score}
\usage{
rank_score(n, method.list = list(name = "Wilcoxon"))
}
\arguments{
\item{n}{A positive integer that specifies the number of units.}

\item{method.list}{A list that specifies the choice of test statistic:
\itemize{
\item \code{list(name = "Wilcoxon")} (the default): the Wilcoxon rank-sum statistic.
\item \code{list(name = "Stephenson", s = 10)}: the Stephenson rank-sum statistic with
parameter \code{s = 10}, where \code{s} is a positive integer.
}}
}
\value{
An \eqn{n \times 1} vector.
}
\description{
\code{rank_score()} calculates the scores of \code{n} units (ordered from low to high)
given a specified test statistic, not allowing for ties.
}
\examples{
rank_score(n = 5, method.list = list(name = "Wilcoxon"))
# c(0.2, 0.4, 0.6, 0.8, 1.0)
rank_score(n = 5, method.list = list(name = "Stephenson", s = 2))
# c(0.00, 0.25, 0.50, 0.75, 1.00)
}
