% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_function_sharp_null.R
\name{rank_treat_relative_to_control}
\alias{rank_treat_relative_to_control}
\title{Calculate the rank of treated units relative to control units}
\usage{
rank_treat_relative_to_control(Z, Y)
}
\arguments{
\item{Z}{Treatment assignment (\eqn{n \times 1} vector).}

\item{Y}{Realized outcome (\eqn{n \times 1} vector, no \code{NA}s).}
}
\value{
A scalar.
}
\description{
Calculate the rank of treated units relative to control units
}
\examples{
Z <- c(1, 0, 0, 1)
Y <- c(1:4)
rank_treat_relative_to_control(Z, Y)
# c(0, 2)
}
